feat: create AuthenticatedHomeContent and GuestHomeContent components for user-specific home page rendering
